Abstract ROSTopicPublisher
Extracted common functionality for asynchronous (threaded) publishing to ROS topics into abstract ROSTopicPublisher class. Also implements a context manager to start and stop publishing.
